backend/requirements.txt
# Travel Itinerary AI Orchestrator - Backend
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Database
sqlalchemy==2.0.36
greenlet>=3.0
asyncpg==0.30.0
alembic==1.14.0
redis==5.2.1

# External APIs (HTTP)
httpx==0.28.1

# LangGraph & AI (Gemini)
langgraph==0.2.60
langchain-google-genai==2.0.8
google-generativeai==0.8.3

# Env
python-dotenv==1.0.1
